+++ conftest.py
@@ -7,7 +7,7 @@ from contextlib import suppress
 import pytest
 
 from panoptes.pocs import hardware
-from panoptes.pocs.config import store as config_store
+from panoptes.utils.config import store as config_store
 from panoptes.pocs.utils.location import download_iers_a_file
 from panoptes.pocs.utils.logger import PanLogger, get_logger