+++ src/panoptes/pocs/mount/__init__.py
@@ -4,7 +4,7 @@ from glob import glob
 from panoptes.utils import error
 from panoptes.utils.library import load_module
 
-from panoptes.pocs.config.store import get_config, set_config
+from panoptes.utils.config.store import get_config, set_config
 from panoptes.pocs.mount.mount import AbstractMount  # noqa
 from panoptes.pocs.utils.location import create_location_from_config
 from panoptes.pocs.utils.logger import get_logger