+++ src/panoptes/pocs/utils/service/power.py
@@ -14,7 +14,7 @@ from typing import Annotated
 from fastapi import Depends, FastAPI, HTTPException, Request
 from pydantic import BaseModel
 
-from panoptes.pocs.config.store import get_config
+from panoptes.utils.config.store import get_config
 from panoptes.pocs.sensor.power import PowerBoard