+++ tests/test_constraints.py
@@ -10,7 +10,7 @@ from panoptes.utils import horizon as horizon_utils
 from panoptes.utils.error import PanError
 from panoptes.utils.serializers import from_yaml
 
-from panoptes.pocs.config.store import get_config
+from panoptes.utils.config.store import get_config
 from panoptes.pocs.scheduler import create_constraints_from_config
 from panoptes.pocs.scheduler.constraint import (
     AlreadyVisited,