+++ tests/test_dome_simulator.py
@@ -1,6 +1,6 @@
 import pytest
 
-from panoptes.pocs.config.store import set_config
+from panoptes.utils.config.store import set_config
 from panoptes.pocs.dome import create_dome_simulator, simulator